requirements.txt
streamlit>=1.28.0
twilio>=8.10.0
sqlalchemy>=2.0.20
python-dotenv>=1.0.0
geopy>=2.3.0
folium>=0.14.0
streamlit-folium>=0.15.1